\name{get.lmer.effects.forBootMer}
\alias{get.lmer.effects.forBootMer}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
   \sQuote{get.lmer.effects} for bootstrap
}
\description{
This is a wrapper for the \sQuote{get.lmer.effects} only intended to prepare bootstrap analyses.
See examples of \sQuote{get.lmer.effects} for further details.
}
\usage{
get.lmer.effects.forBootMer ( lmerObj )
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{lmerObj}{An object of class \sQuote{merMod} or \sQuote{glmerMod}, as created by
  \sQuote{lmer} or \sQuote{glmer} from the \sQuote{lme4} package. }
}
%% \details{}
\value{
A numerical vector of values.
}
%% \references{}
\author{
Sebastian Weirich
}
\examples{
### See examples of 'get.lmer.effects' for further details.
}
