
\name{rater}

\docType{data}

\alias{rater}

\title{Variable ratings for 1287 examinees, 7 variables and 4 raters.}

\description{
This data set contains fictional ratings of achievement test items of 1287 examinees,
7 variables and 4 raters in the long format.
}

\usage{data(rater)}

\format{'data.frame':   4200 obs. of  4 variables
  \describe{
    \item{id}{Student identifier}
    \item{rater}{rater identifier}
    \item{variable}{Variable identifier}
  	\item{value}{The rating given by the corresponding rater to the response of the student to the variable}
 }
}

\source{Simulated data}

%\references{
%}

\keyword{datasets}


