\name{reinsort.col}
\alias{reinsort.col}
\title{Insert Columns into a Data Frame in a Specific Position
}
\description{
  Insert columns into a data frame in specific position
}
\usage{
  reinsort.col(dat, toreinsort, after)
}
\arguments{

  \item{dat}{A data frame}
  \item{toreinsort}{Column name(s) or column number(s) of the colums to be reinserted }
  \item{after}{Column name or column number after which the colums specified in \code{reinsort} should be reinserted.}

}
\details{
}
\value{
  A	data frame with columns in specified positions.
}
