

\name{science1.items}

\docType{data}

\alias{science1.items}

\title{Science achievement test data - Item names}

\description{
This vector contains the names items in data set \code{\link{science1}}
}

%\usage{data(science1.items)}

\format{chr [1:185]}

%\source{}

%\references{
%}

%\keyword{datasets}






