

\name{science1}

\docType{data}

\alias{science1}

\title{Science achievement test data}

\description{
This data set contains responses of 420 students on 185 science items. Additional variables are included:
id, grade, sex, booklet, track, version, and four dummy coded variables that indicate Track x Version groups.
An incomplete block design was used with 4 booklets.
Codes on items are:
"0" - wrong
"1" - right
"mbd" - missing by design
"mbi" - missing by intention
"mir" - missing due to irregular response
}

\usage{data(science1)}

\format{'data.frame':   420 obs. of  195 variables}

\source{Simulated data}

%\references{
%}

\keyword{datasets}






