

\name{science1.scales}

\docType{data}

\alias{science1.scales}

\title{Science achievement test data - Scale definition}

\description{
This data frame contains scale definitions for usage with \code{\link{automateModels}} and data set \code{\link{science1}}
}

%\usage{ }

\format{'data.frame':   185 obs. of  7 variables}

%\source{}

%\references{
%}

\keyword{datasets}






